Move class 'MemoryUsageReport' to be near its usage
