+++ game-core/src/main/java/org/triplea/debug/error/reporting/StackTraceReportModel.java
@@ -8,7 +8,6 @@ import java.util.function.Predicate;
 import javax.annotation.Nonnull;
 import lombok.Builder;
 import org.triplea.debug.LoggerRecord;
-import org.triplea.debug.console.window.MemoryUsageReport;
 import org.triplea.debug.error.reporting.formatting.ErrorReportBodyFormatter;
 import org.triplea.debug.error.reporting.formatting.ErrorReportTitleFormatter;
 import org.triplea.http.client.error.report.ErrorReportRequest;
